+++ crates/oxc_linter/src/utils/react.rs
@@ -8,7 +8,7 @@ use oxc_ast::{
 };
 use oxc_semantic::{AstNode, SymbolFlags};
 
-use crate::{ESLintSettings, JsxA11y, LintContext};
+use crate::{ESLintSettings, LintContext};
 
 pub fn is_create_element_call(call_expr: &CallExpression) -> bool {
     if let Some(member_expr) = call_expr.callee.get_member_expr() {
@@ -225,9 +225,8 @@ pub fn get_element_type(context: &LintContext, element: &JSXOpeningElement) -> O
     };
 
     let ESLintSettings { jsx_a11y, .. } = context.settings();
-    let JsxA11y { polymorphic_prop_name, components } = jsx_a11y;
 
-    if let Some(polymorphic_prop_name_value) = polymorphic_prop_name {
+    if let Some(polymorphic_prop_name_value) = &jsx_a11y.polymorphic_prop_name {
         if let Some(as_tag) = has_jsx_prop_lowercase(element, polymorphic_prop_name_value) {
             if let Some(JSXAttributeValue::StringLiteral(str)) = get_prop_value(as_tag) {
                 return Some(String::from(str.value.as_str()));
@@ -236,7 +235,7 @@ pub fn get_element_type(context: &LintContext, element: &JSXOpeningElement) -> O
     }
 
     let element_type = ident.name.as_str();
-    if let Some(val) = components.get(element_type) {
+    if let Some(val) = jsx_a11y.components.get(element_type) {
         return Some(String::from(val));
     }
     Some(String::from(element_type))
